[skip build] testing travis script

+++ util/travis_compiled_push.sh
@@ -26,21 +26,25 @@ else
 	echo "No essential files modified."
 fi
 
-make ergodox-ez AUTOGEN=true
+if [[ "$TRAVIS_COMMIT_MESSAGE" != *"[skip build]"* ]] ; then
 
-find . -name ".build" | xargs rm -rf
-cd ..
-git clone https://$GH_TOKEN@github.com/jackhumbert/qmk.fm.git
-cd qmk.fm
-git submodule update --init --recursive
-rm -rf keyboard
-rm -rf keyboards
-cp -r ../qmk_firmware/keyboards .
-mkdir keyboards/ergodox_ez/
-cp ../qmk_firmware/util/ergodox_ez.html keyboards/ergodox_ez/index.html
-cp ../qmk_firmware/readme.md qmk_readme.md
-./generate.sh
+	make ergodox-ez AUTOGEN=true
 
-git add -A
-git commit -m "generated from qmk/qmk_firmware@${rev}" 
-git push
+	find . -name ".build" | xargs rm -rf
+	cd ..
+	git clone https://$GH_TOKEN@github.com/jackhumbert/qmk.fm.git
+	cd qmk.fm
+	git submodule update --init --recursive
+	rm -rf keyboard
+	rm -rf keyboards
+	cp -r ../qmk_firmware/keyboards .
+	mkdir keyboards/ergodox_ez/
+	cp ../qmk_firmware/util/ergodox_ez.html keyboards/ergodox_ez/index.html
+	cp ../qmk_firmware/readme.md qmk_readme.md
+	./generate.sh
+
+	git add -A
+	git commit -m "generated from qmk/qmk_firmware@${rev}" 
+	git push
+
+fi
